back/requirements.txt
fastapi
uvicorn[standard]
spotipy
requests
colorthief
python-dotenv
